feat: Adds metadata builder which can generate a commit payload
